Extracted common configuration of E2E tests
