+++ booking-app/src/e2e/java/io/github/marcopaglio/booking/app/swing/MongoBookingSwingAppE2E.java
@@ -12,36 +12,18 @@ import static org.bson.codecs.configuration.CodecRegistries.fromProviders;
 import static org.bson.codecs.configuration.CodecRegistries.fromRegistries;
 import static org.bson.codecs.pojo.Conventions.ANNOTATION_CONVENTION;
 import static org.bson.codecs.pojo.Conventions.USE_GETTERS_FOR_SETTERS;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.swing.launcher.ApplicationLauncher.application;
-import static org.assertj.swing.timing.Pause.pause;
-import static org.assertj.swing.timing.Timeout.timeout;
 
 import java.time.LocalDate;
 import java.util.Arrays;
 import java.util.UUID;
-import java.util.regex.Pattern;
 
-import javax.swing.JFrame;
-
-import org.assertj.swing.annotation.GUITest;
-import org.assertj.swing.core.GenericTypeMatcher;
-import org.assertj.swing.core.matcher.JButtonMatcher;
-import org.assertj.swing.finder.WindowFinder;
-import org.assertj.swing.fixture.FrameFixture;
-import org.assertj.swing.fixture.JButtonFixture;
-import org.assertj.swing.fixture.JLabelFixture;
-import org.assertj.swing.fixture.JListFixture;
-import org.assertj.swing.fixture.JTextComponentFixture;
 import org.assertj.swing.junit.runner.GUITestRunner;
-import org.assertj.swing.junit.testcase.AssertJSwingJUnitTestCase;
-import org.assertj.swing.timing.Condition;
 import org.bson.codecs.configuration.CodecProvider;
 import org.bson.codecs.configuration.CodecRegistry;
 import org.bson.codecs.pojo.PojoCodecProvider;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
-import org.junit.Test;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.runner.RunWith;
 
@@ -61,38 +43,8 @@ import io.github.marcopaglio.booking.model.Reservation;
 
 @DisplayName("End-to-end tests for BookingSwingApp usign MongoDB")
 @RunWith(GUITestRunner.class)
-public class MongoBookingSwingAppE2E extends AssertJSwingJUnitTestCase {
+public class MongoBookingSwingAppE2E extends BookingSwingAppE2E {
 	private static final String DBMS = "MONGO";
-	private static final int TIMEOUT = 5000;
-	private static final int DEFAULT_NUM_OF_CLIENTS = 1;
-	private static final int DEFAULT_NUM_OF_RESERVATIONS = 1;
-
-	private static final String A_FIRSTNAME = "Mario";
-	private static final String A_LASTNAME = "Rossi";
-	private static final UUID A_CLIENT_UUID = UUID.fromString("b8a88ad6-739e-4df5-b3ea-82832a56843a");
-	private static final String ANOTHER_FIRSTNAME = "Maria";
-	private static final String ANOTHER_LASTNAME = "De Lucia";
-
-	private static final String INVALID_FIRSTNAME = "Mari4";
-	private static final String INVALID_LASTNAME = "De_Lucia";
-
-	private static final String A_YEAR = "2022";
-	private static final String A_MONTH = "04";
-	private static final String A_DAY = "24";
-	private static final String A_DATE = A_YEAR + "-" + A_MONTH + "-" + A_DAY;
-	private static final UUID A_RESERVATION_UUID = UUID.fromString("6de5f28a-541a-4699-8377-e0eaa9d2b13e");
-	private static final String ANOTHER_YEAR = "2023";
-	private static final String ANOTHER_MONTH = "09";
-	private static final String ANOTHER_DAY = "05";
-	private static final String ANOTHER_DATE = ANOTHER_YEAR + "-" + ANOTHER_MONTH + "-" + ANOTHER_DAY;
-
-	private static final String INVALID_YEAR = "2O23";
-
-	// regex
-	private static final String NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX =
-			".*" + A_FIRSTNAME + ".*" + A_LASTNAME + ".*" + "|" +
-			".*" + A_LASTNAME + ".*" + A_FIRSTNAME + ".*";
-	private static final String DATE_REGEX = ".*" + A_DATE + ".*";
 
 	private static String mongoHost = System.getProperty("mongo.host", "localhost");
 	private static int mongoPort = Integer.parseInt(System.getProperty("mongo.port", "27017"));
@@ -101,73 +53,6 @@ public class MongoBookingSwingAppE2E extends AssertJSwingJUnitTestCase {
 	private static MongoDatabase database;
 	private static MongoCollection<Client> clientCollection;
 	private static MongoCollection<Reservation> reservationCollection;
-
-	private FrameFixture window;
-
-	private JTextComponentFixture nameFormTxt;
-	private JTextComponentFixture surnameFormTxt;
-	private JTextComponentFixture yearFormTxt;
-	private JTextComponentFixture monthFormTxt;
-	private JTextComponentFixture dayFormTxt;
-	private JLabelFixture formErrorMsgLbl;
-	private JLabelFixture operationErrorMsgLbl;
-	private JButtonFixture addReservationBtn;
-	private JButtonFixture addClientBtn;
-	private JButtonFixture renameBtn;
-	private JButtonFixture rescheduleBtn;
-	private JButtonFixture removeClientBtn;
-	private JButtonFixture removeReservationBtn;
-	private JListFixture clientList;
-	private JListFixture reservationList;
-
-	// pause conditions
-	private Condition untilClientListContainsDifferentNumberOfClientsThanTheDefaultOnes = new Condition(
-			"Client list to contain different number of clients than the default ones") {
-		@Override
-		public boolean test() {
-			return clientList.contents().length != DEFAULT_NUM_OF_CLIENTS;
-		}
-	};
-
-	private Condition untilNameFormsAreReset = new Condition("Name forms to be reset") {
-		@Override
-		public boolean test() {
-			return nameFormTxt.text().isEmpty() && 
-					surnameFormTxt.text().isEmpty();
-		}
-	};
-
-	private Condition untilReservationListContainsDifferentNumberOfReservationThanTheDefaultOnes = new Condition(
-			"Reservation list to contain different number of reservations than the default ones") {
-		@Override
-		public boolean test() {
-			return reservationList.contents().length != DEFAULT_NUM_OF_RESERVATIONS;
-		}
-	};
-
-	private Condition untilDateFormsAreReset = new Condition("Date forms to be reset") {
-		@Override
-		public boolean test() {
-			return yearFormTxt.text().isEmpty() && 
-					monthFormTxt.text().isEmpty() && 
-					dayFormTxt.text().isEmpty();
-		}
-	};
-
-	private Condition untilFormErrorContainsAMessage = new Condition("Form error to contain a message") {
-		@Override
-		public boolean test() {
-			return !formErrorMsgLbl.text().isBlank();
-		}
-	};
-
-	private Condition untilOperationErrorContainsAMessage = new Condition("Operation error to contain a message") {
-		@Override
-		public boolean test() {
-			return !operationErrorMsgLbl.text().isBlank();
-		}
-	};
-
 
 	@BeforeClass
 	public static void setupClient() throws Exception {
@@ -215,36 +100,7 @@ public class MongoBookingSwingAppE2E extends AssertJSwingJUnitTestCase {
 					"--port=" + mongoPort
 			).start();
 		
-		// get a reference of its JFrame
-		window = WindowFinder.findFrame(new GenericTypeMatcher<JFrame>(JFrame.class) {
-			@Override
-			protected boolean isMatching(JFrame frame) {
-				return "BookingApp".equals(frame.getTitle()) && frame.isShowing();
-			}
-		}).using(robot());
-		
-		// text fields
-		nameFormTxt = window.textBox("nameFormTxt");
-		surnameFormTxt = window.textBox("surnameFormTxt");
-		yearFormTxt = window.textBox("yearFormTxt");
-		monthFormTxt = window.textBox("monthFormTxt");
-		dayFormTxt = window.textBox("dayFormTxt");
-		
-		// error labels
-		formErrorMsgLbl = window.label("formErrorMsgLbl");
-		operationErrorMsgLbl = window.label("operationErrorMsgLbl");
-		
-		// buttons
-		addReservationBtn = window.button(JButtonMatcher.withText("Add Reservation"));
-		addClientBtn = window.button(JButtonMatcher.withText("Add Client"));
-		renameBtn = window.button(JButtonMatcher.withText("Rename"));
-		rescheduleBtn = window.button(JButtonMatcher.withText("Reschedule"));
-		removeClientBtn = window.button(JButtonMatcher.withText("Remove Client"));
-		removeReservationBtn = window.button(JButtonMatcher.withText("Remove Reservation"));
-		
-		// lists
-		clientList = window.list("clientList");
-		reservationList = window.list("reservationList");
+		super.onSetUp();
 	}
 
 	@AfterClass
@@ -252,351 +108,8 @@ public class MongoBookingSwingAppE2E extends AssertJSwingJUnitTestCase {
 		mongoClient.close();
 	}
 
-
-	@Test @GUITest
-	@DisplayName("Application is started")
-	public void testApplicationWhenIsStartedShouldShowAllDatabaseElements() {
-		assertThat(clientList.contents())
-			.anySatisfy(e -> assertThat(e).contains(A_FIRSTNAME, A_LASTNAME));
-		
-		assertThat(reservationList.contents())
-			.anySatisfy(e -> assertThat(e).contains(A_DATE));
-	}
-
-
-	////////////// Add client
-	@Test @GUITest
-	@DisplayName("Adding client succeeds")
-	public void testAddClientWhenIsSuccessfulShouldShowTheNewClient() {
-		nameFormTxt.enterText(ANOTHER_FIRSTNAME);
-		surnameFormTxt.enterText(ANOTHER_LASTNAME);
-		
-		addClientBtn.click();
-		
-		pause(untilClientListContainsDifferentNumberOfClientsThanTheDefaultOnes, timeout(TIMEOUT));
-		
-		assertThat(clientList.contents())
-			.anySatisfy(e -> assertThat(e).contains(ANOTHER_FIRSTNAME, ANOTHER_LASTNAME));
-	}
-
-	@Test @GUITest
-	@DisplayName("Name is invalid")
-	public void testAddClientWhenNameIsInvalidShouldShowFormError() {
-		nameFormTxt.enterText(INVALID_FIRSTNAME);
-		surnameFormTxt.enterText(ANOTHER_LASTNAME);
-		
-		addClientBtn.click();
-		
-		pause(untilFormErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(formErrorMsgLbl.text()).contains(INVALID_FIRSTNAME);
-	}
-
-	@Test @GUITest
-	@DisplayName("Surname is invalid")
-	public void testAddClientWhenSurnameIsInvalidShouldShowFormError() {
-		nameFormTxt.enterText(ANOTHER_FIRSTNAME);
-		surnameFormTxt.enterText(INVALID_LASTNAME);
-		
-		addClientBtn.click();
-		
-		pause(untilFormErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(formErrorMsgLbl.text()).contains(INVALID_LASTNAME);
-	}
-
-	@Test @GUITest
-	@DisplayName("Adding client fails")
-	public void testAddClientWhenIsAFailureShouldShowAnOperationError() {
-		nameFormTxt.enterText(A_FIRSTNAME);
-		surnameFormTxt.enterText(A_LASTNAME);
-		
-		addClientBtn.click();
-		
-		pause(untilOperationErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(operationErrorMsgLbl.text()).contains(A_FIRSTNAME, A_LASTNAME);
-	}
-	////////////// Add client
-
-
-	////////////// Rename client
-	@Test @GUITest
-	@DisplayName("Renaming client succeeds")
-	public void testRenameClientWhenIsSuccessfulShouldShowChanges() {
-		clientList.selectItem(Pattern.compile(NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX));
-		nameFormTxt.enterText(ANOTHER_FIRSTNAME);
-		surnameFormTxt.enterText(ANOTHER_LASTNAME);
-		
-		renameBtn.click();
-		
-		pause(untilNameFormsAreReset, timeout(TIMEOUT));
-		
-		assertThat(clientList.contents())
-			.allSatisfy(e -> assertThat(e).doesNotContain(A_FIRSTNAME, A_LASTNAME))
-			.anySatisfy(e -> assertThat(e).contains(ANOTHER_FIRSTNAME, ANOTHER_LASTNAME));
-	}
-
-	@Test @GUITest
-	@DisplayName("New name is invalid")
-	public void testRenameClientWhenNewNameIsInvalidShouldShowFormError() {
-		clientList.selectItem(Pattern.compile(NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX));
-		nameFormTxt.enterText(INVALID_FIRSTNAME);
-		surnameFormTxt.enterText(ANOTHER_LASTNAME);
-		
-		renameBtn.click();
-		
-		pause(untilFormErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(formErrorMsgLbl.text()).contains(INVALID_FIRSTNAME);
-	}
-
-	@Test @GUITest
-	@DisplayName("New surname is invalid")
-	public void testRenameClientWhenNewSurnameIsInvalidShouldShowFormError() {
-		clientList.selectItem(Pattern.compile(NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX));
-		nameFormTxt.enterText(ANOTHER_FIRSTNAME);
-		surnameFormTxt.enterText(INVALID_LASTNAME);
-		
-		renameBtn.click();
-		
-		pause(untilFormErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(formErrorMsgLbl.text()).contains(INVALID_LASTNAME);
-	}
-
-	@Test @GUITest
-	@DisplayName("Renamed client would be homonymic")
-	public void testRenameClientWhenThereIsHomonymyShouldShowAnOperationError() {
-		UUID anotherClientUUID = UUID.fromString("62d3c7b4-b519-46da-ab3a-4934238323b9");
-		addTestClientToDatabase(ANOTHER_FIRSTNAME, ANOTHER_LASTNAME, anotherClientUUID);
-		
-		clientList.selectItem(Pattern.compile(NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX));
-		nameFormTxt.enterText(ANOTHER_FIRSTNAME);
-		surnameFormTxt.enterText(ANOTHER_LASTNAME);
-		
-		renameBtn.click();
-		
-		pause(untilOperationErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(operationErrorMsgLbl.text()).contains(ANOTHER_FIRSTNAME, ANOTHER_LASTNAME);
-	}
-
-	@Test @GUITest
-	@DisplayName("Renaming client fails")
-	public void testRenameClientWhenIsAFailureShouldShowAnOperationError() {
-		clientList.selectItem(Pattern.compile(NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX));
-		nameFormTxt.enterText(ANOTHER_FIRSTNAME);
-		surnameFormTxt.enterText(ANOTHER_LASTNAME);
-		
-		removeTestClientFromDatabase(A_FIRSTNAME, A_LASTNAME);
-		
-		renameBtn.click();
-		
-		pause(untilOperationErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(operationErrorMsgLbl.text()).contains(A_FIRSTNAME, A_LASTNAME);
-	}
-	////////////// Rename client
-
-
-	////////////// Remove client
-	@Test @GUITest
-	@DisplayName("Removing client succeeds")
-	public void testRemoveClientWhenIsSuccessfulShouldMakeDisappearTheClient() {
-		clientList.selectItem(Pattern.compile(NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX));
-		
-		removeClientBtn.click();
-		
-		pause(untilClientListContainsDifferentNumberOfClientsThanTheDefaultOnes, timeout(TIMEOUT));
-		
-		assertThat(clientList.contents())
-			.allSatisfy(e -> assertThat(e).doesNotContain(A_FIRSTNAME, A_LASTNAME));
-	}
-
-	@Test @GUITest
-	@DisplayName("Removing client fails")
-	public void testRemoveClientWhenIsAFailureShouldShowOperationError() {
-		clientList.selectItem(Pattern.compile(NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX));
-		
-		removeTestClientFromDatabase(A_FIRSTNAME, A_LASTNAME);
-		
-		removeClientBtn.click();
-		
-		pause(untilOperationErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(operationErrorMsgLbl.text()).contains(A_FIRSTNAME, A_LASTNAME);
-	}
-	////////////// Remove client
-
-
-	////////////// Add reservation
-	@Test @GUITest
-	@DisplayName("Adding reservation succeeds")
-	public void testAddReservationWhenIsSuccessfulShouldShowTheNewReservation() {
-		clientList.selectItem(Pattern.compile(NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX));
-		yearFormTxt.enterText(ANOTHER_YEAR);
-		monthFormTxt.enterText(ANOTHER_MONTH);
-		dayFormTxt.enterText(ANOTHER_DAY);
-		
-		addReservationBtn.click();
-		
-		pause(untilReservationListContainsDifferentNumberOfReservationThanTheDefaultOnes, timeout(TIMEOUT));
-		
-		assertThat(reservationList.contents())
-			.anySatisfy(e -> assertThat(e).contains(ANOTHER_DATE));
-	}
-
-	@Test @GUITest
-	@DisplayName("Date is invalid")
-	public void testAddReservationWhenDateIsInvalidShouldShowFormError() {
-		clientList.selectItem(Pattern.compile(NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX));
-		yearFormTxt.enterText(INVALID_YEAR);
-		monthFormTxt.enterText(ANOTHER_MONTH);
-		dayFormTxt.enterText(ANOTHER_DAY);
-		
-		addReservationBtn.click();
-		
-		pause(untilFormErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(formErrorMsgLbl.text()).contains(INVALID_YEAR);
-	}
-
-	@Test @GUITest
-	@DisplayName("Selected client is wrong")
-	public void testAddReservationWhenSelectedClientIsWrongShouldShowOperationError() {
-		clientList.selectItem(Pattern.compile(NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX));
-		yearFormTxt.enterText(ANOTHER_YEAR);
-		monthFormTxt.enterText(ANOTHER_MONTH);
-		dayFormTxt.enterText(ANOTHER_DAY);
-		
-		removeTestClientFromDatabase(A_FIRSTNAME, A_LASTNAME);
-		
-		addReservationBtn.click();
-		
-		pause(untilOperationErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(operationErrorMsgLbl.text()).contains(A_FIRSTNAME, A_LASTNAME);
-	}
-
-	@Test @GUITest
-	@DisplayName("Adding reservation fails")
-	public void testAddReservationWhenIsAFailureShouldShowAnOperationError() {
-		clientList.selectItem(Pattern.compile(NAME_THEN_SURNAME_OR_SURNAME_THEN_NAME_REGEX));
-		yearFormTxt.enterText(A_YEAR);
-		monthFormTxt.enterText(A_MONTH);
-		dayFormTxt.enterText(A_DAY);
-		
-		addReservationBtn.click();
-		
-		pause(untilOperationErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(operationErrorMsgLbl.text()).contains(A_DATE);
-	}
-	////////////// Add reservation
-
-
-	////////////// Reschedule reservation
-	@Test @GUITest
-	@DisplayName("Rescheduling reservation succeeds")
-	public void testRescheduleReservationWhenIsSuccessfulShouldShowChanges() {
-		reservationList.selectItem(Pattern.compile(DATE_REGEX));
-		yearFormTxt.enterText(ANOTHER_YEAR);
-		monthFormTxt.enterText(ANOTHER_MONTH);
-		dayFormTxt.enterText(ANOTHER_DAY);
-		
-		rescheduleBtn.click();
-		
-		pause(untilDateFormsAreReset, timeout(TIMEOUT));
-		
-		assertThat(reservationList.contents())
-			.allSatisfy(e -> assertThat(e).doesNotContain(A_DATE))
-			.anySatisfy(e -> assertThat(e).contains(ANOTHER_DATE));
-	}
-
-	@Test @GUITest
-	@DisplayName("New date is invalid")
-	public void testRescheduleReservationWhenNewDateIsInvalidShouldShowFormError() {
-		reservationList.selectItem(Pattern.compile(DATE_REGEX));
-		yearFormTxt.enterText(INVALID_YEAR);
-		monthFormTxt.enterText(ANOTHER_MONTH);
-		dayFormTxt.enterText(ANOTHER_DAY);
-		
-		rescheduleBtn.click();
-		
-		pause(untilFormErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(formErrorMsgLbl.text()).contains(INVALID_YEAR);
-	}
-
-	@Test @GUITest
-	@DisplayName("Rescheduled reservation would be simultaneous")
-	public void testRescheduleReservationWhenThereIsSimultaneityShouldShowAnOperationError() {
-		UUID anotherReservationUUID = UUID.fromString("173d3cc5-4bc9-4815-8cf5-117fb5036909");
-		addTestReservationToDatabase(A_CLIENT_UUID, ANOTHER_DATE, anotherReservationUUID);
-		
-		reservationList.selectItem(Pattern.compile(DATE_REGEX));
-		yearFormTxt.enterText(ANOTHER_YEAR);
-		monthFormTxt.enterText(ANOTHER_MONTH);
-		dayFormTxt.enterText(ANOTHER_DAY);
-		
-		rescheduleBtn.click();
-		
-		pause(untilOperationErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(operationErrorMsgLbl.text()).contains(ANOTHER_DATE);
-	}
-
-	@Test @GUITest
-	@DisplayName("Rescheduling reservation fails")
-	public void testRescheduleReservationWhenIsAFailureShouldShowAnOperationError() {
-		reservationList.selectItem(Pattern.compile(DATE_REGEX));
-		yearFormTxt.enterText(ANOTHER_YEAR);
-		monthFormTxt.enterText(ANOTHER_MONTH);
-		dayFormTxt.enterText(ANOTHER_DAY);
-		
-		removeTestReservationFromDatabase(A_DATE);
-		
-		rescheduleBtn.click();
-		
-		pause(untilOperationErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(operationErrorMsgLbl.text()).contains(A_DATE);
-	}
-	////////////// Reschedule reservation
-
-
-	////////////// Remove reservation
-	@Test @GUITest
-	@DisplayName("Removing reservation succeeds")
-	public void testRemoveReservationWhenIsSuccessfulShouldMakeDisappearTheReservation() {
-		reservationList.selectItem(Pattern.compile(DATE_REGEX));
-		
-		removeReservationBtn.click();
-		
-		pause(untilReservationListContainsDifferentNumberOfReservationThanTheDefaultOnes, timeout(TIMEOUT));
-		
-		assertThat(reservationList.contents())
-			.allSatisfy(e -> assertThat(e).doesNotContain(A_DATE));
-	}
-
-	@Test @GUITest
-	@DisplayName("Removing reservation fails")
-	public void testRemoveReservationWhenIsAFailureShouldShowOperationError() {
-		reservationList.selectItem(Pattern.compile(DATE_REGEX));
-		
-		removeTestReservationFromDatabase(A_DATE);
-		
-		removeReservationBtn.click();
-		
-		pause(untilOperationErrorContainsAMessage, timeout(TIMEOUT));
-		
-		assertThat(operationErrorMsgLbl.text()).contains(A_DATE);
-	}
-	////////////// Remove reservation
-
-
-	public void addTestClientToDatabase(String name, String surname, UUID id) {
+	@Override
+	protected void addTestClientToDatabase(String name, String surname, UUID id) {
 		Client client = new Client(name, surname);
 		client.setId(id);
 		ClientSession session = mongoClient.startSession();
@@ -606,7 +119,8 @@ public class MongoBookingSwingAppE2E extends AssertJSwingJUnitTestCase {
 		session.close();
 	}
 
-	private void removeTestClientFromDatabase(String name, String surname) {
+	@Override
+	protected void removeTestClientFromDatabase(String name, String surname) {
 		ClientSession session = mongoClient.startSession();
 		clientCollection.deleteOne(session, Filters.and(
 						Filters.eq(FIRSTNAME_DB, name),
@@ -614,7 +128,8 @@ public class MongoBookingSwingAppE2E extends AssertJSwingJUnitTestCase {
 		session.close();
 	}
 
-	public void addTestReservationToDatabase(UUID clientId, String date, UUID id) {
+	@Override
+	protected void addTestReservationToDatabase(UUID clientId, String date, UUID id) {
 		Reservation reservation = new Reservation(clientId,  LocalDate.parse(date));
 		reservation.setId(id);
 		ClientSession session = mongoClient.startSession();
@@ -624,7 +139,8 @@ public class MongoBookingSwingAppE2E extends AssertJSwingJUnitTestCase {
 		session.close();
 	}
 
-	private void removeTestReservationFromDatabase(String date) {
+	@Override
+	protected void removeTestReservationFromDatabase(String date) {
 		ClientSession session = mongoClient.startSession();
 		reservationCollection.deleteOne(session, Filters.eq(DATE_DB, LocalDate.parse(date)));
 		session.close();
